fix / improve matching messages
Currently the regex is not using the multiline option. Therefore the `^` results in only matching the first message. Passing the flag `m` allows to match `^` at every line and therefore match all messages.

Additionally the string after the line number and before the newline (replaced by a pipe) is not relevant for the message. It only contains information about the scope but since the message will be shown at a specific location in the file it is sufficient only show the part starting with `D00x something is wrong`.